remove config file, add hooks
