+++ app/sqlcell.py
@@ -10,59 +10,8 @@ import pandas as pd
 import pickle
 from db.engine import EngineHandler, DBSessionHandler
 from args.line_args import ArgHandler
+from hooks.hooks import HookHandler
 
-            
-class HookHandler(DBSessionHandler):
-    """input common queries to remember with a key/value pair. ie, 
-       %%sql hook
-       \d=<common query>"
-       \dt=<another common query>"""
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-        
-    def is_engine(self, engine: str):
-        try:
-            create_engine(engine)
-            return True
-        except:
-            return False
-        
-    def add(self, line, cell):
-        "add hook to db"
-        cmds_to_add = []
-        hooks = cell.split('\n\n')
-        for hook in hooks:
-            hook = hook.strip()
-            if hook:
-                key_engine, cmd = [i.strip() for i in hook.split('=', 1)]
-                key, engine = key_engine.split(' ')
-                engine = engine
-                cmds_to_add.append((key, engine, cmd))
-
-        for key, engine, cmd in cmds_to_add:
-            engine = self.db_info.get(engine, engine)
-            is_engine = self.is_engine(engine)
-            if not is_engine:
-                raise Exception('Alias not found or engine argument error')
-            self.session.add(self.Hooks(key=key, engine=engine, cmd=cmd))
-        self.session.commit()
-        return self
-        
-    def run(self, cell, engine_var):
-        cell = cell.replace('~', '').split(' ')
-        cell, cmd_args = cell[0], cell[1:]
-        hook_query = self.session.query(self.Hooks).filter_by(key=cell).first()
-        hook_cmd = hook_query.cmd
-        if engine_var:
-            hook_engine = create_engine(str(engine.url))
-        else:
-            hook_engine = create_engine(hook_query.engine)
-        SQLCell.current_hook_engine = hook_engine
-        return hook_engine, hook_cmd.format(*cmd_args)
-    
-    def refresh(self, cell):
-        self.session.query(Hooks).delete()
-        self.session.commit()
 
 @magics_class
 class SQLCell(Magics, DBSessionHandler):
@@ -70,7 +19,7 @@ class SQLCell(Magics, DBSessionHandler):
     current_engine = False
     current_hook_engine = False
     modes = ['query', 'hook', 'refresh']
-    # consider cfg file for these types of params:
+    # consider yaml file for these types of params:
     hook_indicator = '~'
     
     def __init__(self, shell, data):
@@ -161,13 +110,14 @@ class SQLCell(Magics, DBSessionHandler):
         
         engine = self.get_engine(engine_var)
         ########################## HookHandler logic ########################
-        hook_handler = HookHandler()
+        hook_handler = HookHandler(engine)
         if hook:
             hook_handler.add(line, cell)
             return ('Hook successfully registered')
         
         if cell.startswith(self.hook_indicator):
             engine, cell = hook_handler.run(cell, engine_var)
+            SQLCell.current_hook_engine = hook_handler.hook_engine
         ########################## End HookHandler logic ####################
         ############################ Refresh logic ##########################
         if refresh and cell in self.refresh_optinos:
